# document/swagger-document/Company_api_doc.json
{"swagger":"2.0","info":{"description":"Stock Market Company REST API","version":"V2","title":"Company REST API","contact":{"name":"Jack.L","url":"https://github.com/liuchang3373/stock-market","email":"lchangcd@cn.ibm.com"},"license":{"name":"The MIT License","url":"https://opensource.org/licenses/MIT"}},"host":"localhost:8081","basePath":"/","tags":[{"name":"ipo-controller","description":"IPO Controller"},{"name":"company-controller","description":"Company Controller"}],"paths":{"/company/create":{"post":{"tags":["company-controller"],"summary":"createCompany","operationId":"createCompanyUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"companyDto","description":"companyDto","required":true,"schema":{"$ref":"#/definitions/CompanyDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CompanyResponse"}}}}},"/company/delete":{"delete":{"tags":["company-controller"],"summary":"deleteCompany","operationId":"deleteCompanyUsingDELETE","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"companyId","in":"query","description":"companyId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BaseResponse"}}}}},"/company/getCompanyByCode":{"get":{"tags":["company-controller"],"summary":"getCompany","operationId":"getCompanyUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"companyCode","in":"query","description":"companyCode","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CompanyResponse"}}}}},"/company/ipo/create":{"post":{"tags":["ipo-controller"],"summary":"createIPO","operationId":"createIPOUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"ipoDto","description":"ipoDto","required":true,"schema":{"$ref":"#/definitions/IPODto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/IPOResponse"}}}}},"/company/ipo/delete":{"delete":{"tags":["ipo-controller"],"summary":"deleteIPO","operationId":"deleteIPOUsingDELETE","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"ipoId","in":"query","description":"ipoId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BaseResponse"}}}}},"/company/ipo/getIPOById":{"get":{"tags":["ipo-controller"],"summary":"getIPO","operationId":"getIPOUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"query","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/IPOResponse"}}}}},"/company/ipo/listAllIPOs":{"get":{"tags":["ipo-controller"],"summary":"getAllIPOs","operationId":"getAllIPOsUsingGET","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/IPOListResponse"}}}}},"/company/ipo/update":{"put":{"tags":["ipo-controller"],"summary":"updateIPO","operationId":"updateIPOUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"IPODto","description":"IPODto","required":true,"schema":{"$ref":"#/definitions/IPODto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/IPOResponse"}}}}},"/company/listAllCompanies":{"get":{"tags":["company-controller"],"summary":"getAllCompanies","operationId":"getAllCompaniesUsingGET","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CompanyListResponse"}}}}},"/company/update":{"put":{"tags":["company-controller"],"summary":"updateCompany","operationId":"updateCompanyUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"companyDto","description":"companyDto","required":true,"schema":{"$ref":"#/definitions/CompanyDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CompanyResponse"}}}}}},"definitions":{"BaseResponse":{"type":"object","properties":{"code":{"type":"string","enum":["SUCCESS","FAILURE","UN_AUTHORIZED","NOT_FOUND","MSG_NOT_READABLE","METHOD_NOT_SUPPORTED","MEDIA_TYPE_NOT_SUPPORTED","REQ_REJECT","INTERNAL_SERVER_ERROR","PARAM_MISS","PARAM_TYPE_ERROR","PARAM_BIND_ERROR","PARAM_VALID_ERROR"]},"message":{"type":"string"},"success":{"type":"boolean"}}},"CompanyDto":{"type":"object","properties":{"boardDirector":{"type":"string"},"brief":{"type":"string"},"ceoName":{"type":"string"},"companyCode":{"type":"string"},"companyName":{"type":"string"},"exchangeId":{"type":"integer","format":"int32"},"exchangeName":{"type":"string"},"id":{"type":"integer","format":"int32"},"sectorId":{"type":"integer","format":"int32"},"sectorName":{"type":"string"},"turnOver":{"type":"number","format":"float"}}},"CompanyListResponse":{"type":"object","properties":{"code":{"type":"string","enum":["SUCCESS","FAILURE","UN_AUTHORIZED","NOT_FOUND","MSG_NOT_READABLE","METHOD_NOT_SUPPORTED","MEDIA_TYPE_NOT_SUPPORTED","REQ_REJECT","INTERNAL_SERVER_ERROR","PARAM_MISS","PARAM_TYPE_ERROR","PARAM_BIND_ERROR","PARAM_VALID_ERROR"]},"companyListDto":{"type":"array","items":{"$ref":"#/definitions/CompanyDto"}},"message":{"type":"string"},"success":{"type":"boolean"}}},"CompanyResponse":{"type":"object","properties":{"code":{"type":"string","enum":["SUCCESS","FAILURE","UN_AUTHORIZED","NOT_FOUND","MSG_NOT_READABLE","METHOD_NOT_SUPPORTED","MEDIA_TYPE_NOT_SUPPORTED","REQ_REJECT","INTERNAL_SERVER_ERROR","PARAM_MISS","PARAM_TYPE_ERROR","PARAM_BIND_ERROR","PARAM_VALID_ERROR"]},"companyDto":{"$ref":"#/definitions/CompanyDto"},"message":{"type":"string"},"success":{"type":"boolean"}}},"IPODto":{"type":"object","properties":{"companyName":{"type":"string"},"exchangeId":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int32"},"openDateTime":{"type":"string","format":"date"},"pricePerShare":{"type":"number","format":"float"},"remarks":{"type":"string"},"sectorId":{"type":"integer","format":"int32"},"totalNumberShare":{"type":"number","format":"float"}}},"IPOListResponse":{"type":"object","properties":{"code":{"type":"string","enum":["SUCCESS","FAILURE","UN_AUTHORIZED","NOT_FOUND","MSG_NOT_READABLE","METHOD_NOT_SUPPORTED","MEDIA_TYPE_NOT_SUPPORTED","REQ_REJECT","INTERNAL_SERVER_ERROR","PARAM_MISS","PARAM_TYPE_ERROR","PARAM_BIND_ERROR","PARAM_VALID_ERROR"]},"ipoListDto":{"type":"array","items":{"$ref":"#/definitions/IPODto"}},"message":{"type":"string"},"success":{"type":"boolean"}}},"IPOResponse":{"type":"object","properties":{"code":{"type":"string","enum":["SUCCESS","FAILURE","UN_AUTHORIZED","NOT_FOUND","MSG_NOT_READABLE","METHOD_NOT_SUPPORTED","MEDIA_TYPE_NOT_SUPPORTED","REQ_REJECT","INTERNAL_SERVER_ERROR","PARAM_MISS","PARAM_TYPE_ERROR","PARAM_BIND_ERROR","PARAM_VALID_ERROR"]},"ipoDetailDto":{"$ref":"#/definitions/IPODto"},"message":{"type":"string"},"success":{"type":"boolean"}}}}}